rename locales json files
