+++ src/plugins/i18n/locales/de.json
